Move users to api
